requirements.txt
streamlit
PyPDF2
google-generativeai
requests
reportlab
typing-extensions
io
json
grpcio
protobuf
pillow
